Finished refactoring. Ready for the unit tests.
