--- DFC.App.JobProfileOverview.Tests.IntegrationTests.API/Support/CommonAction/CommonAction_General.cs
@@ -1,26 +1,0 @@
-﻿using DFC.App.JobProfileOverview.Tests.IntegrationTests.API.Support.Interface;
-using Newtonsoft.Json;
-using System;
-using System.Linq;
-using System.Text;
-
-namespace DFC.App.JobProfileOverview.Tests.IntegrationTests.API.Support
-{
-    internal partial class CommonAction : IGeneralSupport
-    {
-        private static readonly Random Random = new Random();
-
-        public string RandomString(int length)
-        {
-            const string chars = "ABCDEFGHIJKLMNOPQRSTUVWXYZ";
-            return new string(Enumerable.Repeat(chars, length)
-              .Select(s => s[Random.Next(s.Length)]).ToArray());
-        }
-
-        public byte[] ConvertObjectToByteArray(object obj)
-        {
-            string serialisedContent = JsonConvert.SerializeObject(obj);
-            return Encoding.ASCII.GetBytes(serialisedContent);
-        }
-    }
-}
